Library delegates: reduce includes, remove unneeded functions and debug asserts
